feat:접근성 font 속성 추가 및 접근성 메뉴 컴포넌트화
* font
   - 색상 : color
   - 크기 : size
* AccessibilityMenu.tsx 추가
